+++ cms/delProduct.php
@@ -1,17 +1,25 @@
 <?php
-require "include/init.php";
-require "include/topBar.php";
-require "include/sideBar.php";
+require_once "include/init.php";
+include "include/topBar.php";
+include "include/sideBar.php";
+
+if(!isset($_SESSION["userRole"]) || $_SESSION["userRole"] < 1) {
+	http_response_code(403);
+	echo "403 Forbidden";
+	exit;
+}
 
 if(isset($_GET["confirm"]) && isset($_GET["productID"]) && is_numeric($_GET["productID"])){
 	$stmt = $dbh->prepare("DELETE FROM product WHERE productID = :productID");
 	$stmt->bindParam(":productID", $_GET["productID"]);
 	$stmt->execute();
+
+	// This is safe because is_numeric() returned true
 	unlink("../img/products/" . $_GET["productID"] . "-large.jpg");
 	unlink("../img/products/" . $_GET["productID"] . "-small.jpg");
+
 	echo "Product is verwijderd";
 }
-
 else if(isset($_GET["productID"]) && is_numeric($_GET["productID"])) {
 	$stmt = $dbh->prepare("SELECT name FROM product WHERE productID = :productID");
 	$stmt->bindParam(":productID", $_GET["productID"]);
@@ -24,11 +32,9 @@ else if(isset($_GET["productID"]) && is_numeric($_GET["productID"])) {
 	}
 
 	$result = $stmt->fetch();
-	print("Weet u zeker dat u <i>" . $result["name"] . "</i> wilt verwijderen<br>");
-	print("<a class='btn btn-primary' href=delProduct.php?productID=" . $_GET["productID"] . "&confirm=1>Ja</a>");
-	print("<a class='btn btn-primary' href=products.php>Nee</a>");
-
-	// TODO: delete images
+	echo "Weet u zeker dat u <i>" . $result["name"] . "</i> wilt verwijderen<br>";
+	echo "<a class='btn btn-primary' href=delProduct.php?productID=" . $_GET["productID"] . "&confirm=1>Ja</a>";
+	echo "<a class='btn btn-primary' href=products.php>Nee</a>";
 }
 else {
 	echo "Wat doe jij nou";
